Add fastpath for space in V_SRB2PgenericDrawString
NFNT032 as of 1.3.5 doesn't exist. Currently, the engine relies on W_CachePatchName2 to return NULL when it fails to find NFNT032, which under certain situations it does not. As a hack, we can check if this lumpname appears, then set the patch variable to NULL ourselves. I would assume this saves computation time for handling spaces, since orignally we would have to search every single lump in every single wad to just throw NULL, it does add extra computational time for every other character, so improvements to this check could be made.