Implement actual PROPER way of handling 10000+ freeslots
The crash I discovered was due to a string read overflow on freeslots with an index greater than 10000. The freeslots name table has a strict limit of 7 characters, presuming the 7th is a nullbyte. A new case for values past 10000(and non crash wise, to 99999) was made in order to preserve the 6 character limit and also keep us (hopefully) well away from the limit for a while.